refactored code again
