+++ app.js
@@ -7,10 +7,12 @@ const app = express();
 const initializePassport = require('./src/middleware/configPassport')
 const bodyParser = require('body-parser');
 const path = require('path');
-const {errorMiddleware} = require('./middleware/error');
+const {errorMiddleware} = require('./src/middleware/error');
 // const flash = require("express-flash");
 
+const communityRouter = require('./src/routes/community_routes');
 const userRouter = require('./src/routes/auth_routes');
+const tripRouter = require('./src/routes/trip_routes');
 
 //Additional middlewares
 app.use(express.json())
@@ -43,7 +45,9 @@ app.use(express.urlencoded({ extended: false }));
 // app.set("view engine", "ejs");
 
 //Router
-app.use("/users", userRouter);
+app.use( "/", userRouter);
+app.use('/trips',tripRouter);
+app.use('/communities',communityRouter)
 
 
 // Funtion inside passport which initializes passport
